build: Move libopensaf_core back to libdir and update README [#2298]
An application binary linked with the OpenSAF AIS libraries may fail to load
after upgrading to OpenSAF 5.2. The reason is that libopensaf_core.so.0 has
moved from /usr/local/lib to /usr/local/lib/opensaf. This problem can happen if
the following two conditions are met:

* The application was linked without using the -Wl,--as-needed option (and this
  option is not enabled by default by the Linux distribution used when building
  the application binary)
* The directory /usr/local/lib/opensaf is not listed in LD_LIBRARY_PATH or
  /etc/ld.so.conf

This is fixed by:

* Moving libopensaf_core back to /usr/local/lib
* Updating the documentation to mention that -Wl,--as-needed must to be used
  when linking with the OpenSAF libraries, to avoid similar problems in the
  future.