amfd: fix default value of saAmfSIPrefActiveAssignments [1190]
AMF spec specifies two different default values for saAmfSIPrefActiveAssignments in:
1)section 3.6.5: Default is  preferred number of assigned service
  units i.e saAmfSGNumPrefAssignedSUs.
2)section 8.11: default value as 1.
Currently if a user does not configure any value for this attribute, then
IMM sets default value to 1. Because of this when SG is unlocked or lots
of SUs are unlocked, AMF assigns SI to only one SU. For other SUs to get
assignment for this SI, a CCB modification is needed. This causes problem
in scaling up application as CCB modification needs to be performed each time
a SU is added.

This enhancement also honours default value given in section 3.6.5. So the default
value of saAmfSIPrefActiveAssignments will be either saAmfSGNumPrefAssignedSUs or 1
based on user choice.
Following are conditions in which different default values will be honoured:
-if a user configures saAmfSIPrefActiveAssignments=1 then SI will assigned to only one
SU.This is to ensure backward compatibility.
-if a user does not configure attribute saAmfSIPrefActiveAssignments in application or
 deletes this attributes via CCB operation then AMFD will still keep default value. This
 is also to ensure backward compatibility.
-if a user sets saAmfSIPrefActiveAssignments=0 via CCB or in application conf then AMFD will
 use section 3.6.5 definition for default value i.e saAmfSGNumPrefAssignedSUs.