imm: Sync latest ccb-id to sync clients [#2323]
When finalizing sync, immnd coordinator stores latest ccb-id in ImmsvCcbOutcomeList.
It is stored as a ccb with ccb-id being zero.
The sync clients update mLatestCcbId when receiving ND2ND_SYNC_FINALIZE_2 message.

When rolling upgrading, it's not a valid case when an old-versioned node joins the cluster.
After rebooting, the nodes are supposed to be upgraded and run with new version.
We don't have to worry about old-versioned sync-client receiving ccb-id = 0.

We don't have to sync latest admo-id and latest implementer-id because
amdo and implementer are all "cleared" in IMMND when headless.
There won't be any problem if IMMD reset IMMD_CB.admo_id_count and IMMD_CB.impl_count.