smf: cli-command does not wait for nodes to start [#1969]
While doing the wrapup of a one-step upgrade with reboot the other
controller takes extra time to come back up. Now when we have a CLI
command in the campaign doing a operation on that missing controller the
campaign will fail. The reason is that there is no timer to wait for
node up. This patch uses a prexisting function with a timer that waits
until a node comes up for 10 minutes. The time is taken from SMF
configuration 'smfRebootTimeout'.