log: fix checkpoint dest_names in open stream request [#2434]
Handling of checkpoint for stream open is in serveral places. Handling of checkpoint
is called in proc_stream_open_msg  forgot to add checkpoint of destination name.

Refactor so that handling of checkpoint data for stream open is done in one place
and destination name was already added in checkpoint data at this place