amfd: skip unassigned SUSIs in SU for updating saAmfSUAssignedSIs [#2496]
In the reported problem, AMFD crashes while trying to free invalid memory.
AMFD is trying to update value of saAmfSUAssignedSIs to IMM.

SU assignments counters saAmfSUNumCurrActiveSIs and saAmfSUNumCurrStandbySIs
reflects only active and standby HA state. If a SI is in quiesced state, it
is still assigned to this SU and su->list_of_susi will not be null. But
saAmfSUNumCurrActiveSIs will be 0.
Thus in the reported problem, no memory allocation was done
and AMFD will crash while freeing invalid memory.

Patch counts all those SUSIs which are not in unassigned fsm state in SU.
A SUSI with fsm state unassigned is getting deleted and thus not counted.