amfnd: retry on ERR_NOT_EXIST [#2490]
On a congested network, sometimes a newly created IMM object can take some
time to be available on other nodes.

In our test, a new SU is created on SC-1 and unlocked. But sometimes
it fails on a remote node due to:

2017-05-19 13:55:19 SC-2 osafamfnd[258]: ER amf_saImmOmSearchInitialize_o2 failed: 12

To get around this, we will retry on SA_AIS_ERR_NOT_EXIST a few times.