rde: Avoid sending messages to peers that are not up [#2488]
Replace the request-reply message exchange between RDE peers with a simpler
protocol that just sends the reply message just after receiving peer up
notifications. This will minimize the probability of trying to send to a peer
that is down. Also remove the retransmission logic that was mostly needed
because messages can be received before the node up event, which is now no
longer an issue.