dtmd: send broadcast messages after initial discovery [#2522]
If TCP closes a socket (due to TCP_USER_TIMEOUT or some other issue) because
of a network problem, and DTM declares that the node has exited the cluster,
the node never attempts to come back into the cluster if it does not reboot.

Initial discovery broadcasts stop after 5 seconds.

Discovery broadcasts should continue after the initial discovery phase at a
reduced rate.