dtm: Re-factor the inter-node receive logic [#2554]
The DTM receive logic was unnecessarily complex: it called recv() in many
different places and had conditional statements that executed different pieces
of code depending on how many bytes that had been received on the stream socket
so far. The code has now been refactored so that it uses only one single recv()
statement and one single check to determine if it has received a full message or
not.