imm: remove disconnected appliers [#2579]
When an applier is disconnected, it will be removed from the system after time period set in minApplierTimeout attribute in IMM object.
If the time is set to 0, it will work as it works today, and it will never be removed from the system.

The time set in minApplierTimeout attribute guarantees that the applier will be "alive" for at least the time set in the attribute.
When this period expires, with the next clean the basement call, the applier will be removed.

The time set in minApplierTimeout attribute is in seconds.

To enable this feature, protocol51710 must be enabled and minApplierTimeout attribute value must be great than 0.