amfnd: fix opensaf shutdown and active monitoring failure [#2493]
If active monitoring command is wrongly configured by the user, AMF reports fault
on NPI component and AMFND restarts it. Since every time active monitoring command
fails, component is getting continuously faulted. When OpenSAF is stopped on this node
then amfnd asserted.

In the reported issue, continuous fault and recovery of component is not getting
escalated to node-failover as compRestartProb is very low. During OpenSAF shutdown,
amfnd does not terminate component as a part removal of assignment and deletes ssui.
In component clean up phase, amfnd terminates component but does not find si for it and
asserts.

Patch ensures that AMFND terminates component as a part of removal of assignment during
OpenSAF stop.