log: fix incorrect handling of version when initializing OpenSAF APIs [#2516]
The version used when initializing the API is in many cases stored in a global
variable and this global variable is used every time the API is initialized.
The version is given as a pointer to this variable. The problem is that this
variable is defined as an [in/out] parameter that gives the version to use when
initializing [in] but the value shall be changed by the agent to the highest
version supported by the initialized service [out]. This means that the next
time the API is initialized it may be initialized with a higher version.

This patch help to fix that always initialize OpenSaf API with correct version
in log service.