smf: try to wait for opensafd status before reboot [#2464]
There are cases when opensafd startup is still ongoing and SMF will send
out a reboot command for a node. Because opensafd has taken a lock the
reboot command will not be able to call opensafd stop. It is suggested
that SMF tries to wait for the release of the lock with "opensafd
status". The waiting time is short and SMF continues with reboot even if
the lock is not released.

- Change smf remote command to have two versions, one that logs errors of
  the endpoint command and one without error logging.