smf: execLevel for balanced upgrades [#2555]
Previously, balanced procedures got the highest execLevel and were
executed last in the chain. There are cases where it is needed to
execute procedures after the balanced procedures are completed. To offer
this feature SMF can check which execLevel existing procedures that are
used for balanced upgrade have. Inserting the balanced procedure one
step after the highest of those procedures and moving existing
procedures that interfere to a higher execLevel. This allows to
configure procedures that are not part of the balanced group to be
executed after a balanced procedure.