dtm: Extract Multicast class for node discovery message handling [#991]
Extract a class with the name Multicast, that handles send and receive of DTM
node discovery datagram messages. The underlying protocol can be IP broadcast,
IP multicast, and - implemented in this ticket: IP unicast.

Note that this refactoring isn't complete - the new class has private methods
that are in fact more or less unmodified copies of the old C functions.