dtm: Support multiple OpenSAF internal log streams [#2165]
Extend the OpenSAF internal log service so that it supports multiple log streams
in addition to the MDS log. All log streams are received on the same socket and
the stream name is extracted from the MSGID field. Each stream is written to a
separate file based on the name of the stream.