ckptnd: fix crash during checkpoint open timeout with large sections [#1510]
ckptnd crashes

When opening a collocated checkpoint replica where the active has large
numbers of sections (~200k), the sync from the active can timeout with
errorcode SA_AIS_ERR_TRY_AGAIN. In this case the code deletes the memory
for the node, but does not delete the node from the db. When the
checkpoint access is tried again, the freed memory for the node is still
in the db, and ckptnd crashes.

Delete the node from the db if the node is deleted during the open.