nid: Derive node ID from TIPC address when not managing TIPC [#2653]
If OpenSAF is not configured to manage TIPC, we don't need to require the
presence of /etc/opensaf/slot_id or /var/lib/opensaf/node_id. Instead, we create
the file /var/lib/opensaf/node_id ourselves based on the TIPC address of the
node we are running on. The same translation formula is used (in the opposite
direction) as when OpenSAF is managing TIPC.