plm: fix PLM to run on OpenSUSE Leap 42.2 [#2480]
plmd fails to execute functions in libplms_hpi.so after it is dlopened and
dlsym'd.

HPI is not linked against libplms_hpi.so, and because the library is
dlopened with LAZY, trying to execute any saHpiXXX call fails to resolve
the symbol. PLM then exits. There are also some multiply defined symbols.

Link libplms_hpi.so against the HPI lib. And make variables declared in
header files "extern".