plm: terminate child EEs when parent is terminated [#2572]
If the hypervisor EE is terminated by shutting down plmcd on the host, the
child EEs (VMs) are not terminated and put into Uninstantiated state. Then
when plmcd is restarted on the host, the child VMs are not started or
restarted because PLM thinks they are instantiated.

Logic is wrong in plms_plmc_tcp_disconnect_process when checking for child
EEs.

Terminate child EEs when the host is terminated.