msgd: fix double delete of node info [#2712]
msgd can crash if multiple nodes in the cluster are rebooted at the same time.

"is_node_down" and "is_clm_down" can both be set to true, if the CLM tracking
callback comes in quickly. If this occurs during middleware failover, the
cleanup code in msgd can cause a crash because it tries to delete the same
node twice.

Don't delete the same node twice.