plmd: fix mbc in PLM [#2724]
MBC isn't working in PLM, so no info is being checkpointed to the standby plmd.

When the code to handle more than 2 SCs was put in to PLM, the MBC selection
object was gotten at a later time -- after the while loop containing the "poll"
system call. Thus, the mbc file descriptor was never being set in the poll call.

Move the setting of the mbc file descriptor to inside the while loop, so it gets
set.