clmd: add dynamically created EEs to PLM entity group on standby [#2730]
If EEs and corresponding CLM nodes are dynamically created, after a middleware
si-swap when the former standby has become active, then one of those EEs is
rebooted, clmd has not enabled PLM readiness state tracking on the EE and will
not know when it comes back. Thus, the node will not be allowed back into the
cluster because it thinks it is not a member.

The dynamically created EE is not being added to the PLM entity group on the
standby.

Add the dynamically created EE to the PLM entity group on the standby.