plm: handle plmc clients which abruptly terminated [#2529]
In virtual environments nodes can reboot very quickly (less than 1 minute). If
the reboot is abrupt, plmd may not be aware that the EE went down until after
it has already come back up because plmd relies on the TCP connection to plmcd
on the node. In this case, plmd will set the readiness state to OOS after the
EE is already back up. This causes CLM to evict the node from the cluster. plmd
should use TCP_USER_TIMEOUT to notice that plmcd has exited abruptly.

This enhancement also refactors the threading involved with handling the plm
clients, to support a large number of them.