imm: fix immnd restarted during SYNC finalize [#2749]
The call to saImmOiImplementerSet and saImmOiImplementerClear are allowed during
sync.  Therefore, there is possibility the messages of these calls arrived at
the sync-client after finalizeSync was sent from IMMND coord but before
finalizeSync arrived at sync-client (over fevs).

In that case, the implementer record stored in sImplementerVector database will
then be different with the implementer record in finalizeSync and this mismatch
caused the IMMND sync-client restart with current implementation.

This ticket adds a check to ignore that implementer record in finalizeSync
as it is obsolete instead of terminating the sync-client.