plm: don't set readiness state to in-service if EE is terminating [#2734]
If an EE goes down during a controller switchover the TERMINATED message
sent by plmc to plmd may not be received because of the switch over.

In this case the EE will be stuck in terminating presence state.

If any parent of the EE is in OOS, then we can definitely set the presence
state to UNINSTANTIATED after the switchover. If not, then we can just set
the management-lost flag because we don't know whether or not the EE
terminated.