plm: handle race condition for EE instantiation [#2514]
Child EE which is a controller can get shutdown because its parent EE (host)
has not connected to PLM, yet.

If the controller is a VM, and the host is a payload, there is a race
condition when instantiating the EEs. If the host doesn't connect to PLM
first, then when the controller EE (child of host EE) connects to PLM, it
see that the host isn't instantiated, and shuts itself down.

If the controller child EE instantiates before the host has connected to PLM,
set a 20 second timer. If the host doesn't instantiate within this time, then
all child EEs will be shut down.