amfnd: Check UNASSIGNED CSI in csi assigment removal [#2741]
A csi can be marked as UNASSIGNED from a su restart or this csi
is newly added from scaling out. This unassigned csi currently
makes the csi assignment removal sequence running non-stop, since
the removal sequence is not checking UNASSIGNED state.

Patch considers UNASSIGNED state of csi to be ignored so the
removal sequence can move on