imm: immnd coredump during SYNC finalize [#2775]
The call to saImmOiImplementerSet and saImmOiImplementerClear are allowed during
sync.  Therefore, there is possibility the messages of these calls arrived at
veteran nodes after finalizeSync was sent from IMMND coord but before
finalizeSync arrived at the veterans (over fevs).

In that case, the implementer record stored in sImplementerVector database will
then be different with the implementer record in finalizeSync and this mismatch
caused the IMMND veterans aborted with current implementation.

This ticket adds a check to ignore that implementer record in finalizeSync
as it is obsolete instead of terminating the veterans.