ntfd: Cold sync reader to the standby ntfd after rebooting the standby controller [#2757]
Assumpt that the reader information is updated to the standby ntfd via checkpoint
upon reception of reader APIs requests. However, if the standby controller reboots
and comes up, the standby ntfd still has none of readers information which is
available at the active ntfd. Now if a switchover happens, the new active will not
be able to process the reader APIs requests with existing reader handles.

This patch adds reader information as part of cold sync