dtm: Re-try binding to handle IPv6 DAD race condition [#2596]
When using IPv6, we must re-try bind() to handle the potential IPv6 DAD
(duplicate address detection) race condition.

After an IPv6 address has been added to an interface, there will be a short
period when the address is considered tentative and bind() will fail with
EADDRNOTAVAIL.

This commit also fixes a race in between the creation of the TCP listening
socket and the node discovery broadcast, by moving the socket creation to the
main thread.