amfnd: Discard new assignment while su is under failover [#2773]
When two errors happen to component and both escalates to a su
failover. For the first su failover recovery, amfd will send
new assignment to su. However, the second error happens just
before the su receives new assignment. Currently, amfnd creates
a susi record for the new assignment of first recovery, but amfnd
will not issue csi callback to component since component is being
failed. When the new assignment of second recovery comes, amfnd
finds that the susi record had been created, and not process this
new assignment. At the end, component is not getting any csi
callback in both recovery.

This patch makes amfnd not to process the first new assignment
while su is being failed over. This case is similar to the assignment
removal is being discarded while su is under failover.