nid: Set TasksMax systemd option to infinity for opensafd service [#2788]
When using systemd version 227 or later, set the option TasksMax for OpenSAF to
inifinity. This option specifies the maximum number of processes and threads
within the control group for the OpenSAF service. The reason why OpenSAF needs a
large number of processes and threads is that all application processes started
by AMF are running within the same control group, unless they explicitly move
out of it.