msgnd: prevent race condition during q transfer [#2816]
During q transfer when new node is opening the q, msgnd fails to create the
runtime IMM object for the queue, and the open fails.

When the transfer is done, the old side and owner of the runtime object doesn't
delete the IMM object until after the q transfer response is sent. This is a
race condition. If the new side tries to create the runtime object before the
old side has deleted it, the opening of the queue on the new side fails.

Delete the runtime object before sending the q transfer response.