plmd: connect to hypervisor after middleware switchover [#2817]
Any PLM admin operation which requires hypervisor assistance (e.g. unlock-in,
abrupt restart) will fail after middleware switchover.

When plmcds are reconnecting to the new active plmd, the plmd does not attempt
to connect to the hypervisor if the EE is a virtual machine monitor.

Connect to the hypervisor when the virtual machine monitor EE reconnects, and
perform any admin-pending-operations that occurred while the hypervisor was
out of contact.