plmd: handle admin-operation-pending for EE unlock [#2819]
If EE unlock fails, it is never retried when management is regained. The EE
just sits in LOCKED admin state.

If EE unlock fails, the code continues as if it did succeed, setting readiness
state to in-service, etc.

If EE unlock fails, just return ERR_DEPLOYMENT immediately, and don't set
anything else.