imm: improve cascade delete [#2667]
When an object is deleted, and the object has children, the delete meesage is
sent for each deleted object to PBE.  Since there are a lot of messages in the
cascade delete from IMMND to PBE at once, there is a limitation that the cascade
delete should not be done on object that contains more than 10000 object.  More
than 10000 object may cause buffer overload (e.g. TIPC_ERR_OVERLOAD), and
messages might be lost.

The improvement should send only one message to PBE which will contain only the
root object. The rest of cascade delete will be on PBE side.