imma: Do not finalize previously instialized privateOmHandle in saImmOiAugmentCcbInitialize [#2827]
Currently in saImmOiAugmentCcbInitialize, if OI augmented ccb handle
initialization request does not result in OK or TRY_AGAIN, the
privateOmHandle will be finalized and the private OM will be destroyed.

That is fine if the privateOmHandle is initialized in the current
session. But if the privateOmHandle was already initialized in another
previous callback, which means the privateOmHandle record was added to
imma_oi_ccb_record of the corresponding client node, finalizing this
privateOmHandle without cleaning-up imma_oi_ccb_record may call up
unexpected behavior of the Implementer.

This patch checks if the privateOmHandle is the previous initialized
privateOmHandle to avoid finalizing it. This privateOmHandle will be
automatically finalized and cleanup-up in CCB abort or CCB apply.