mds: Update test suite 14 to avoid perpetually hanging [#2831]
Currently in suite 14, mdstest sends messages with MDS_SENDTYPE_SNDACK
and MDS_SENDTYPE_SNDRSP types to the same destination, this sometimes
makes the Response thread to catch the wrong event and does not respond
to MDS_SENDTYPE_SNDRSP call. This causes the test case to hang forever.

This patch changes the destination of MDS_SENDTYPE_SNDRSP call to
NCSMDS_SVC_ID_INTERNAL_MIN to avoid the above issue.