imm: make version parameter in immutil_xxx non-const [#2830]
The version in saImmO{m,i}Initialize is input/output parameter and is declared
as non-constant for both IMM OM and OI API according to SAF spec.
But in immutil wrapper library, some are declared as constant parameter
and the version is not updated before returning from such wrappers.

This patch makes that parameter non-const [in/out] and do update the version
before returning from wrapper APIs. Also fix the wrong usages of
these wrappers - passed constant version, in some services/applications.