dtm: Add --delete option to osaflog command for deleting log streams [#2837]
Make it possible to delete log streams in the internal OpenSAF log server.  This
will free up resources for log streams that are no longer used, as well as make
it possible to create a new log stream with the same name but different
configuration options (max file size and number of backups).