base: Re-factor the timer implementation [#2440]
Re-factor the NCS timer implementation and provide the AIS TMR API as an
alternative way to use the timer implementation, alongside with the old NCS
API. The AIS TMR API is intended for internal use in OpenSAF services for now,
but it could be exported as an official OpenSAF API later if it turns out to be
working well. OpenSAF services using the TMR API instead of the NCS API will no
longer need a dedicated timer thread.

The new timer implementation is using a C++ STL multiset for the timer queue,
and a Linux timerfd for the selection object.